+++ 2014Spring/src/main/res/layout/activity_main.xml
@@ -25,6 +25,6 @@
         android:layout_width="@dimen/navigation_drawer_width"
         android:layout_height="match_parent"
         android:layout_gravity="start"
-        android:name="org.techbooster.app.abc.NavigationDrawerFragment" />
+        android:name="org.techbooster.app.abc.fragments.NavigationDrawerFragment" />
 
 </android.support.v4.widget.DrawerLayout>
